+++ gui/integrator.h
@@ -8,15 +8,15 @@ namespace gui {
 
 class Integrator : public QObject {
   Q_OBJECT
-  Q_PROPERTY(float from MEMBER from_)
-  Q_PROPERTY(float to MEMBER to_)
-  Q_PROPERTY(float step MEMBER step_)
-  Q_PROPERTY(float answer READ answer NOTIFY integrated)
+  Q_PROPERTY(double from MEMBER from_)
+  Q_PROPERTY(double to MEMBER to_)
+  Q_PROPERTY(double step MEMBER step_)
+  Q_PROPERTY(double answer READ answer NOTIFY integrated)
 
  public:
   explicit Integrator(QObject* parent = 0);
 
-  float answer() { return answer_; }
+  double answer() { return answer_; }
 
  signals:
   void integrated();
@@ -26,9 +26,9 @@ class Integrator : public QObject {
 
  private:
   eval::Integrator integrator_;
-  float from_;
-  float to_;
-  float step_;
-  float answer_;
+  double from_;
+  double to_;
+  double step_;
+  double answer_;
 };
 }
